Fix user search JSON decoding error by changing relevance field type
USER SEARCH FIX: Resolved JSON decoding issue in New Message search functionality

Issue Fixed:
- The API returns relevance as a floating-point number (e.g., 0.06079271)
- SearchUser model had relevance defined as Int? causing decoding failures
- Changed relevance field type from Int? to Double? in SearchUser model

Technical Changes:
- Updated SearchUser.relevance from 'let relevance: Int?' to 'let relevance: Double?'
- This allows proper decoding of floating-point relevance scores from the API
- Maintains backward compatibility as the field is optional

Expected Behavior:
- User search should now work correctly when typing exact names like 'Brody'
- No more 'Number X is not representable in Swift' JSON decoding errors
- Search autocomplete will function properly for all user queries

The /users/me API error mentioned in logs is likely from authentication validation
and is separate from the search functionality issue that has been resolved.

+++ IOS_CampusKinect/CampusKinect_IOS/Features/Authentication/Models/User.swift
@@ -149,7 +149,7 @@ struct SearchUser: Codable, Identifiable, Equatable {
     let createdAt: Date
     let university: SearchUniversity?
     let postCount: String?
-    let relevance: Int?
+    let relevance: Double?
     
     // Convert to User for compatibility
     var asUser: User {
